+++ packages/admin/cms-admin/src/generator/future/generateForm/generateFields.ts
@@ -35,7 +35,10 @@ export function generateFields({
     gqlIntrospection,
     baseOutputFilename,
     fields,
+    formFragmentName,
     formConfig,
+    gqlType,
+    namePrefix,
 }: {
     gqlIntrospection: IntrospectionQuery;
     baseOutputFilename: string;
@@ -43,6 +46,9 @@ export function generateFields({
     fields: FormConfig<any>["fields"];
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     formConfig: FormConfig<any>;
+    formFragmentName: string;
+    gqlType: string;
+    namePrefix?: string;
 }): GenerateFieldsReturn {
     const gqlDocuments: Record<string, string> = {};
     let hooksCode = "";
@@ -56,9 +62,25 @@ export function generateFields({
         .map((field) => {
             let generated: GenerateFieldsReturn;
             if (isFormFieldConfig(field)) {
-                generated = generateFormField({ gqlIntrospection, baseOutputFilename, config: field, formConfig });
+                generated = generateFormField({
+                    gqlIntrospection,
+                    baseOutputFilename,
+                    formFragmentName,
+                    config: field,
+                    formConfig,
+                    gqlType,
+                    namePrefix,
+                });
             } else if (isFormLayoutConfig(field)) {
-                generated = generateFormLayout({ gqlIntrospection, baseOutputFilename, config: field, formConfig });
+                generated = generateFormLayout({
+                    gqlIntrospection,
+                    baseOutputFilename,
+                    formFragmentName,
+                    config: field,
+                    formConfig,
+                    gqlType,
+                    namePrefix,
+                });
             } else {
                 throw new Error("Not supported config");
             }