+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @comet/eslint-config
 
+## 7.3.2
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.3.2
+
+## 7.3.1
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.3.1
+
+## 7.3.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.3.0
+
 ## 7.2.1
 
 ### Patch Changes